Maximized window swapping no longer disappears.
This fixes the issue where if you window swapped a maximized window, it
would disappear off screen. It is a hack fix in that instead of +7
pixels we +8 pixels for when it is slightly larger. This works on the
snapper window, along with windows like explorer. Not tested against
other windows like Visual Studio or Steam, which the +7 pixels already
did not work appropriately. That will still come in another fix where we
get the proper window bounds.